+++ src/__init__.py
@@ -1,3 +1,3 @@
 """Waffler core modules"""
 
-__version__ = "3.11.2"
+__version__ = "3.11.3"